backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.9
nltk==3.8.1
supabase==2.9.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.0
pytesseract==0.3.10
Pillow==11.1.0
